refactor: split types into entity-specific files and update imports
