+++ src/entities/album/ui/AlbumCard.tsx
@@ -1,10 +1,10 @@
 import { useCallback, type FC } from "react";
-import type { Album } from "../../../shared/model/types";
 import styles from "./AlbumCard.module.css";
 import { useNavigate } from "react-router-dom";
 import { useTheme } from "../../../shared/lib/theme/useTheme";
 import { Button } from "../../../shared/ui/Button/Button";
 import clsx from "clsx";
+import type { Album } from "../model/types";
 
 type AlbumCardProps = {
   album: Album;
