+++ src/entities/photo/api/photosApi.ts
@@ -1,5 +1,5 @@
 import { createApi, fetchBaseQuery } from "@reduxjs/toolkit/query/react";
-import type { Photo } from "../../../shared/model/types";
+import type { Photo } from "../model/types";
 
 export const photosApi = createApi({
   reducerPath: "photosApi",
