+++ src/entities/todo/model/types.ts
@@ -1,0 +1,6 @@
+export type Todo = {
+  userId: number;
+  id: number;
+  title: string;
+  completed: boolean;
+};
