+++ src/widgets/PostList/model/hooks/usePostWithComments.ts
@@ -1,5 +1,8 @@
 import { useGetCommentsByPostIdQuery } from "../../../../entities/comment/api/commentsApi";
-import type { Post, PostWithComments } from "../../../../shared/model/types";
+import type {
+  Post,
+  PostWithComments,
+} from "../../../../entities/post/model/types";
 
 export const usePostWithComments = (post: Post): PostWithComments => {
   const { data: comments = [], error } = useGetCommentsByPostIdQuery(post.id);
